+++ dubbo-samples-context/pom.xml
@@ -11,5 +11,10 @@
 
     <artifactId>dubbo-samples-context</artifactId>
 
-
+    <dependencies>
+        <dependency>
+            <groupId>org.springframework</groupId>
+            <artifactId>spring-context</artifactId>
+        </dependency>
+    </dependencies>
 </project>
