+++ dubbo-samples-context/src/main/java/org/apache/dubbo/samples/context/ContextConsumer.java
@@ -15,11 +15,10 @@
  *   limitations under the License.
  */
 
-package com.alibaba.dubbo.samples.context;
-
-import com.alibaba.dubbo.samples.context.api.ContextService;
+package org.apache.dubbo.samples.context;
 
 import org.apache.dubbo.rpc.RpcContext;
+import org.apache.dubbo.samples.context.api.ContextService;
 import org.springframework.context.support.ClassPathXmlApplicationContext;
 
 public class ContextConsumer {
@@ -32,8 +31,11 @@ public class ContextConsumer {
         String hello = contextService.sayHello("world"); // call remote method
 
         boolean isConsumerSide = RpcContext.getContext().isConsumerSide();
+        System.out.println("isConsumerSide:" + isConsumerSide);
         String application = RpcContext.getContext().getUrl().getParameter("application");
+        System.out.println("application:" + application);
         String serverIP = RpcContext.getContext().getRemoteHost();
+        System.out.println("serverIP:" + serverIP);
 
         System.out.println(hello); // get result
