+++ dubbo-samples-context/src/main/resources/spring/dubbo-context-consumer.xml
@@ -31,6 +31,6 @@
 
     <!-- generate proxy for the remote service, then demoService can be used in the same way as the
     local regular interface -->
-    <dubbo:reference id="demoService" check="false" interface="com.alibaba.dubbo.samples.context.api.ContextService"/>
+    <dubbo:reference id="demoService" check="false" interface="org.apache.dubbo.samples.context.api.ContextService"/>
 
 </beans>
